# backend/swagger.yaml
swagger: "2.0"
info:
  title: "API de Productos"
  description: "Documentación de la API para gestionar productos"
  version: "1.0.0"
host: "localhost:5000"
schemes:
  - "http"
paths:
  /auth/register:
    post:
      summary: "Registrar un nuevo usuario"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              email:
                type: "string"
              password:
                type: "string"
              business_name:
                type: "string"
      responses:
        201:
          description: "Usuario registrado exitosamente"
        400:
          description: "Datos inválidos"
  /auth/login:
    post:
      summary: "Iniciar sesión"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "Inicio de sesión exitoso"
          schema:
            type: "object"
            properties:
              access_token:
                type: "string"
        401:
          description: "Credenciales inválidas"
...
  /auth/reset-password:
    post:
      summary: "Restablecer contraseña"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              token:
                type: "string"
              new_password:
                type: "string"
      responses:
        200:
          description: "Contraseña restablecida exitosamente"
        400:
          description: "Token inválido o datos inválidos"
...